v7
ON/OFF functionality with left button push works(!!) 
changed activate function to only call allLEDs() function, removed the if/else if logic to not repeatedly change the values of the flags
added new function called deactivate to turn off LEDs. 
in adjustbrightness function, put call to allLEDs into an if statement to adjust brightness only if the lights are on.